src/openworldlib/synthesis/visual_generation/solaris/requirements.txt
einops~=0.8.0
flax==0.10.6
numpy==1.26.0
opencv-python-headless==4.11.0.86
requests==2.32.3
tqdm~=4.66.5
torch==2.8.0
torchvision==0.23.0
timm==1.0.15
hydra-core==1.3.2
optax~=0.2.4
orbax-checkpoint==0.11.16
omegaconf~=2.3.0
av==14.2.0
safetensors~=0.4.5
wandb==0.19.8
tensorstore==0.1.71
black==25.1.0
isort==6.0.1
decord==0.6.0

matplotlib==3.10.3
dm-pix==0.4.4
pytest